remove rolap from olap Signed-off-by: dbulahov <bulahovdenis@gmail.com>
