+++ mondrian/src/main/java/org/eclipse/daanse/rolap/aggmatch/jaxb/RegexMapper.java
@@ -11,7 +11,7 @@
  *   SmartCity Jena - initial
  *   Stefan Bischof (bipolis.org) - initial
  */
-package org.eclipse.daanse.olap.rolap.aggmatch.jaxb;
+package org.eclipse.daanse.rolap.aggmatch.jaxb;
 
 import java.util.ArrayList;
 import java.util.List;
