+++ mondrian/src/main/java/org/eclipse/daanse/rolap/aggmatch/jaxb/package-info.java
@@ -14,4 +14,4 @@
 //TODO: RM EXPORT
 @org.osgi.annotation.bundle.Export
 @org.osgi.annotation.versioning.Version("0.0.1")
-package org.eclipse.daanse.olap.rolap.aggmatch.jaxb;
+package org.eclipse.daanse.rolap.aggmatch.jaxb;
