+++ mondrian/src/main/java/org/eclipse/daanse/rolap/core/package-info.java
@@ -14,4 +14,4 @@
 //TODO: RM EXPORT
 @org.osgi.annotation.bundle.Export
 @org.osgi.annotation.versioning.Version("0.0.1")
-package org.eclipse.daanse.olap.rolap.api;
+package org.eclipse.daanse.rolap.core;
