+++ mondrian/src/test/java/mondrian/rolap/aggmatcher/DefaultRuleTest.java
@@ -5,7 +5,22 @@
 * You must accept the terms of that agreement to use this software.
 *
 * Copyright (c) 2002-2017 Hitachi Vantara..  All rights reserved.
-*/
+ *
+ * ---- All changes after Fork in 2023 ------------------------
+ *
+ * Project: Eclipse daanse
+ *
+ * Copyright (c) 2023 Contributors to the Eclipse Foundation.
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors after Fork in 2023:
+ *   SmartCity Jena - initial
+ */
 
 package mondrian.rolap.aggmatcher;
 
@@ -14,19 +29,19 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.util.Iterator;
 import java.util.List;
 
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.AggRule;
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.AggRules;
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.CharCaseEnum;
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.FactCountMatch;
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.FactCountMatchRef;
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.ForeignKeyMatch;
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.LevelMap;
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.LevelMapRef;
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.MeasureMap;
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.MeasureMapRef;
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.Regex;
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.TableMatch;
-import org.eclipse.daanse.olap.rolap.aggmatch.jaxb.TableMatchRef;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.AggRule;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.AggRules;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.CharCaseEnum;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.FactCountMatch;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.FactCountMatchRef;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.ForeignKeyMatch;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.LevelMap;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.LevelMapRef;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.MeasureMap;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.MeasureMapRef;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.Regex;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.TableMatch;
+import org.eclipse.daanse.rolap.aggmatch.jaxb.TableMatchRef;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
