+++ mondrian/src/test/java/org/eclipse/daanse/olap/tests/TestSetup.java
@@ -1,7 +1,20 @@
+/*
+ * Copyright (c) 2025 Contributors to the Eclipse Foundation.
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *   SmartCity Jena, Stefan Bischof - initial
+ *
+ */
 package org.eclipse.daanse.olap.tests;
 
 import org.eclipse.daanse.olap.calc.base.compiler.BaseExpressionCompilerFactory;
-import org.eclipse.daanse.olap.rolap.core.BasicContext;
+import org.eclipse.daanse.rolap.core.BasicContext;
 import org.osgi.test.common.annotation.Property;
 import org.osgi.test.common.annotation.Property.TemplateArgument;
 import org.osgi.test.common.annotation.Property.ValueSource;
@@ -39,18 +52,18 @@ public class TestSetup {
 //            @Property(key = "test.exec", source = ValueSource.TestUniqueId), })
     @interface WithFoodMartSchemaRecord {
     }
-    
-    
+
+
     @WithFactoryConfiguration(factoryPid = BaseExpressionCompilerFactory.PID, location = "?", properties = {
             @Property(key = "test.exec", source = ValueSource.TestUniqueId), })
     @interface WithBaseExpressionCompilerFactory {
     }
-    
+
 //    @WithFactoryConfiguration(factoryPid = JdbcStatisticsProvider.PID, location = "?", properties = {
 //            @Property(key = "test.exec", source = ValueSource.TestUniqueId), })
 //    @interface WithJdbcStatisticsProvider {
 //    }
 
-    
+
 
 }
